+++ backend/config/urls.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
 
+from django.conf import settings
 from django.contrib import admin
 from django.urls import include, path
+from django.conf.urls.static import static
 from drf_spectacular.views import SpectacularAPIView, SpectacularRedocView, SpectacularSwaggerView
 
 urlpatterns = [
@@ -20,4 +22,4 @@ urlpatterns = [
     path("api/", include("chat.urls")),
     path("api/", include("email_organiser.urls")),
     path("api/", include("dashboard.urls")),
-]
+] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)